[ntuple] Move RFieldBase to separate TU
This is the first step to split the large RField.cxx and matches the
RFieldBase.hxx header introduced by commit f62b5ea6ca.
